+++ src/components/Post/PostLayout.jsx
@@ -4,8 +4,8 @@ import styled from "styled-components";
 import { CenterLayoutBorder } from "../../style/Mixin";
 
 import PostLayoutNavbar from "./PostLayoutNavbar";
-import SweetPosting from "../SweetPosting";
-import Post from "../Post";
+import SweetPosting from "../elements/SweetPosting";
+import Post from "../elements/Post";
 
 const PostLayout = () => {
   return (
