--- src/components/Register.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Register = () => {
-  return <div>Register</div>;
-};
-
-export default Register;
