+++ src/pages/BookmarkPage.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import Layout from "../components/Layout/Layout";
 import SideBar from "../components/Layout/SideBar";
 import SearchBar from "../components/Layout/SearchBar";
-import BookmarkLayout from "../components/Layout/BookmarkLayout";
+import BookmarkLayout from "../components/Bookmark/BookmarkLayout";
 
 const BookmarkPage = () => {
   return (
