+++ src/pages/MainPage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import Layout from "../components/Layout/Layout";
 import SideBar from "../components/Layout/SideBar";
-import PostLayout from "../components/Layout/PostLayout";
+import PostLayout from "../components/Post/PostLayout";
 import SearchBar from "../components/Layout/SearchBar";
 
 const MainPage = () => {
