+++ src/pages/ProfilePage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Layout from "../components/Layout/Layout";
-import ProfileLayout from "../components/Layout/ProfileLayout";
+import ProfileLayout from "../components/Profile/ProfileLayout";
 import SearchBar from "../components/Layout/SearchBar";
 import SideBar from "../components/Layout/SideBar";
